requirements.txt
# Core Web Framework
fastapi==0.120.1
uvicorn[standard]==0.32.1

# HTTP Client
httpx==0.28.1

# LangChain Ecosystem
langchain==0.4.1
langchain-community==0.4.1
langchain-core==1.0.1
langchain-openai==1.0.1
langgraph==1.0.1
langsmith==0.4.38

# OpenAI API
openai==2.6.1

# Environment Management
python-dotenv==1.2.1

# Type Support
typing-extensions==4.15.0

# Testing
pytest==8.4.2
pytest-asyncio==0.24.0

# If some dependecny is still accidently missing then just simply install it